Add functionality for tracking 'invisible' coin.
Replace coin generating functions with a single function
`generate_coin(bool: randomize_coin)` that places two coins, one of
which is the real goal and one which is invisible (only there as
placeholder). If randomize_coin is true, then the goal coin is placed
randomly, otherwise it is placed at the end of the level.

Add entry `coinrun_invisible_coin_collected` to the info dict. This is
true for all steps within an episode after the agent 'collects' the
invisible coin by accident.