Replace goal shapes in maze II
This fixes a bug in the maze II experiment. Prior to this, shapes (star
and gem) were barely distinguishable (or not at all, depending on env seed).
This broke the experiment.

In this commit I fix two of three environments, third pending, by
replacing the star shape with a diagonal line.